Fixed the ui issues related to filters in the business and freelancer marketplace.
